requirements.txt
absl-py==2.2.0
addict==2.4.0
astunparse==1.6.3
basicsr==1.4.2
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cv==1.0.0
cycler==0.12.1
facenet-pytorch==2.6.0
facexlib==0.3.0
filelock==3.18.0
filterpy==1.4.5
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2025.3.0
future==1.0.0
gast==0.6.0
gfpgan==1.3.8
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
idna==3.10
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
keras==3.9.0
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.44.0
lmdb==1.6.2
lz4==4.4.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
mtcnn==1.0.0
namex==0.0.8
networkx==3.4.2
numba==0.61.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
pandas==2.2.3
pillow==10.2.0
platformdirs==4.3.7
protobuf==5.29.4
psutil==7.0.0
py-cpuinfo==9.0.0
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
realesrgan==0.3.0
requests==2.32.3
rich==13.9.4
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
sympy==1.13.1
tb-nightly==2.20.0a20250326
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.37.1
tensorflow_cpu==2.19.0
termcolor==2.5.0
threadpoolctl==3.6.0
tifffile==2025.3.13
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
triton==2.2.0
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.96
ultralytics-thop==2.0.14
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
yapf==0.43.0
gunicorn